requirements.txt
# Project homepage https://github.com/3mpr/OSRBox

PySerial
keyboard
py2exe
